Remove old lexer!
