+++ Cargo.toml
@@ -21,7 +21,6 @@ itertools      = "0.8.0"
 lazy_static    = "1.0.0"
 libc           = "0.2.21"
 log            = "0.4.4"
-regex          = "1.0.0"
 target         = "1.0.0"
 tempdir        = "0.3.5"
 unicode-width  = "0.1.3"
